FinalYear-Project/requirements.txt
lxml==4.6.3
tensorflow-cpu==2.5.0
sendgrid==6.7.0
plotly==4.14.3
numpy==1.19.5
Flask==1.1.2
seaborn==0.11.1
pandas==1.2.4
requests==2.25.1
matplotlib==3.2
chart_studio==1.1.0
beautifulsoup4==4.9.3
ipython==7.24.1
scikit_learn==0.24.2
